#2 #4 Extract configuration classes, Comment mock test
