+++ src/main/kotlin/com/gmaur/investment/robotadvisor/RobotAdvisorApp.kt
@@ -1,28 +1,24 @@
 package com.gmaur.investment.robotadvisor
 
-import arrow.core.Either
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
-import com.gmaur.investment.robotadvisor.domain.AssetAllocation
-import com.gmaur.investment.robotadvisor.domain.Portfolio
 import com.gmaur.investment.robotadvisor.domain.PortfolioRebalancer
 import com.gmaur.investment.robotadvisor.infrastructure.OperationMapper
+import com.gmaur.investment.robotadvisor.infrastructure.Rebalance
 import com.gmaur.investment.robotadvisor.infrastructure.RebalanceRequest
 import org.springframework.boot.ApplicationArguments
 import org.springframework.boot.ApplicationRunner
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
-import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Import
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
 
 @Configuration
 @SpringBootApplication
-@EnableAutoConfiguration
-@ComponentScan(basePackages = arrayOf("com.gmaur.investment.robotadvisor"))
+@Import(value = [AppConfiguration::class])
 @RestController
 class RobotAdvisorApp(private val portfolioRebalancer: PortfolioRebalancer) : ApplicationRunner {
 
@@ -57,19 +53,6 @@ class RobotAdvisorApp(private val portfolioRebalancer: PortfolioRebalancer) : Ap
         println("Application running!")
         //TODO AGB healtcheck
     }
-}
-
-data class Rebalance(val current: Portfolio, val ideal: AssetAllocation) {
-
-    companion object {
-        fun parse(rebalanceRequest: RebalanceRequest): Either<List<Error>, Rebalance> {
-            if (rebalanceRequest.current != null && rebalanceRequest.ideal != null) {
-                return Either.Right(Rebalance(current = rebalanceRequest.current, ideal = rebalanceRequest.ideal))
-            }
-            return Either.Left(listOf(Error("null values")))
-        }
-    }
-
 }
 
 fun main(args: Array<String>) {
