+++ src/main/kotlin/com/gmaur/investment/robotadvisor/domain/PortfolioRebalancer.kt
@@ -1,7 +1,7 @@
 package com.gmaur.investment.robotadvisor.domain
 
-open class PortfolioRebalancer(private val fixedStrategy: RebalancingStrategy) {
+open class PortfolioRebalancer(private val strategy: RebalancingStrategy) {
     open fun rebalance(ideal: AssetAllocation, current: Portfolio): Operations {
-        return fixedStrategy.rebalance(ideal, current)
+        return strategy.rebalance(ideal, current)
     }
 }
